+++ autokeras/__init__.py
@@ -1,36 +1,37 @@
 from autokeras.auto_model import AutoModel
 from autokeras.engine.block import Block
-from autokeras.engine.block import Head
-from autokeras.engine.block import Node
-from autokeras.hypermodels.basic import ConvBlock
-from autokeras.hypermodels.basic import DenseBlock
-from autokeras.hypermodels.basic import Embedding
-from autokeras.hypermodels.basic import ResNetBlock
-from autokeras.hypermodels.basic import RNNBlock
-from autokeras.hypermodels.basic import XceptionBlock
-from autokeras.hypermodels.head import ClassificationHead
-from autokeras.hypermodels.head import RegressionHead
-from autokeras.hypermodels.node import ImageInput
-from autokeras.hypermodels.node import Input
-from autokeras.hypermodels.node import StructuredDataInput
-from autokeras.hypermodels.node import TextInput
-from autokeras.hypermodels.preprocessing import FeatureEncoding
-from autokeras.hypermodels.preprocessing import ImageAugmentation
-from autokeras.hypermodels.preprocessing import Normalization
-from autokeras.hypermodels.preprocessing import TextToIntSequence
-from autokeras.hypermodels.preprocessing import TextToNgramVector
-from autokeras.hypermodels.reduction import Merge
-from autokeras.hypermodels.reduction import SpatialReduction
-from autokeras.hypermodels.reduction import TemporalReduction
-from autokeras.hypermodels.wrapper import ImageBlock
-from autokeras.hypermodels.wrapper import StructuredDataBlock
-from autokeras.hypermodels.wrapper import TextBlock
-from autokeras.task import ImageClassifier
-from autokeras.task import ImageRegressor
-from autokeras.task import StructuredDataClassifier
-from autokeras.task import StructuredDataRegressor
-from autokeras.task import TextClassifier
-from autokeras.task import TextRegressor
+from autokeras.engine.head import Head
+from autokeras.engine.node import Node
+from autokeras.hypermodels import ClassificationHead
+from autokeras.hypermodels import ConvBlock
+from autokeras.hypermodels import DenseBlock
+from autokeras.hypermodels import Embedding
+from autokeras.hypermodels import FeatureEncoding
+from autokeras.hypermodels import Flatten
+from autokeras.hypermodels import ImageAugmentation
+from autokeras.hypermodels import ImageBlock
+from autokeras.hypermodels import Merge
+from autokeras.hypermodels import Normalization
+from autokeras.hypermodels import RegressionHead
+from autokeras.hypermodels import ResNetBlock
+from autokeras.hypermodels import RNNBlock
+from autokeras.hypermodels import SpatialReduction
+from autokeras.hypermodels import StructuredDataBlock
+from autokeras.hypermodels import TemporalReduction
+from autokeras.hypermodels import TextBlock
+from autokeras.hypermodels import TextToIntSequence
+from autokeras.hypermodels import TextToNgramVector
+from autokeras.hypermodels import XceptionBlock
+from autokeras.nodes import ImageInput
+from autokeras.nodes import Input
+from autokeras.nodes import StructuredDataInput
+from autokeras.nodes import TextInput
+from autokeras.tasks import ImageClassifier
+from autokeras.tasks import ImageRegressor
+from autokeras.tasks import StructuredDataClassifier
+from autokeras.tasks import StructuredDataRegressor
+from autokeras.tasks import TextClassifier
+from autokeras.tasks import TextRegressor
 
 from .utils import check_tf_version
