+++ autokeras/adapters/__init__.py
@@ -1,0 +1,6 @@
+from autokeras.adapters.input_adapter import ImageInputAdapter
+from autokeras.adapters.input_adapter import InputAdapter
+from autokeras.adapters.input_adapter import StructuredDataInputAdapter
+from autokeras.adapters.input_adapter import TextInputAdapter
+from autokeras.adapters.output_adapter import ClassificationHeadAdapter
+from autokeras.adapters.output_adapter import RegressionHeadAdapter
