+++ autokeras/encoder.py
@@ -1,10 +1,10 @@
 import numpy as np
 import tensorflow as tf
 
-from autokeras.engine import picklable
+from autokeras.engine import serializable
 
 
-class Encoder(picklable.Picklable):
+class Encoder(serializable.Serializable):
     """Base class for encoders of the prediction targets.
 
     # Arguments
@@ -47,15 +47,17 @@ class Encoder(picklable.Picklable):
         raise NotImplementedError
 
     def get_config(self):
-        return {'num_classes': self.num_classes}
+        return {
+            'num_classes': self.num_classes,
+            'labels': self._labels,
+            'int_to_label': self._int_to_label,
+        }
 
-    def get_state(self):
-        return {'labels': self._labels,
-                'int_to_label': self._int_to_label}
-
-    def set_state(self, state):
-        self._labels = state['labels']
-        self._int_to_label = state['int_to_label']
+    @classmethod
+    def from_config(cls, config):
+        obj = super().from_config(config)
+        obj._labels = config['labels']
+        obj._int_to_label = config['int_to_label']
 
 
 class OneHotEncoder(Encoder):
@@ -71,14 +73,15 @@ class OneHotEncoder(Encoder):
         super().__init__(**kwargs)
         self._label_to_vec = {}
 
-    def get_state(self):
-        state = super().get_state()
-        state.update({'label_to_vec': self._label_to_vec})
-        return state
+    def get_config(self):
+        config = super().get_config()
+        config.update({'label_to_vec': self._label_to_vec})
+        return config
 
-    def set_state(self, state):
-        super().set_state(state)
-        self._label_to_vec = state['label_to_vec']
+    @classmethod
+    def from_config(cls, config):
+        obj = super().from_config(config)
+        obj._label_to_vec = config['label_to_vec']
 
     def fit_with_labels(self, data):
         """Create mapping from label to vector, and vector to label.
@@ -152,14 +155,15 @@ class LabelEncoder(Encoder):
         super().__init__(**kwargs)
         self._label_to_int = {}
 
-    def get_state(self):
-        state = super().get_state()
-        state.update({'label_to_int': self._label_to_int})
-        return state
+    def get_config(self):
+        config = super().get_config()
+        config.update({'label_to_int': self._label_to_int})
+        return config
 
-    def set_state(self, state):
-        super().set_state(state)
-        self._label_to_int = state['label_to_int']
+    @classmethod
+    def from_config(cls, config):
+        obj = super().from_config(config)
+        obj._label_to_int = config['label_to_int']
 
     def fit_with_labels(self, data):
         """Fit the encoder with all the labels.
