+++ autokeras/hypermodels/__init__.py
@@ -1,0 +1,20 @@
+from autokeras.hypermodels.basic import ConvBlock
+from autokeras.hypermodels.basic import DenseBlock
+from autokeras.hypermodels.basic import Embedding
+from autokeras.hypermodels.basic import ResNetBlock
+from autokeras.hypermodels.basic import RNNBlock
+from autokeras.hypermodels.basic import XceptionBlock
+from autokeras.hypermodels.heads import ClassificationHead
+from autokeras.hypermodels.heads import RegressionHead
+from autokeras.hypermodels.preprocessing import FeatureEncoding
+from autokeras.hypermodels.preprocessing import ImageAugmentation
+from autokeras.hypermodels.preprocessing import Normalization
+from autokeras.hypermodels.preprocessing import TextToIntSequence
+from autokeras.hypermodels.preprocessing import TextToNgramVector
+from autokeras.hypermodels.reduction import Flatten
+from autokeras.hypermodels.reduction import Merge
+from autokeras.hypermodels.reduction import SpatialReduction
+from autokeras.hypermodels.reduction import TemporalReduction
+from autokeras.hypermodels.wrapper import ImageBlock
+from autokeras.hypermodels.wrapper import StructuredDataBlock
+from autokeras.hypermodels.wrapper import TextBlock
