+++ autokeras/hypermodels/heads.py
@@ -1,16 +1,14 @@
-import pandas as pd
-import tensorflow as tf
 from tensorflow.keras import layers
 from tensorflow.python.util import nest
 
-from autokeras import encoder
+from autokeras import adapters
 from autokeras import keras_layers
 from autokeras import utils
-from autokeras.engine import block as block_module
+from autokeras.engine import head as head_module
 from autokeras.hypermodels import reduction
 
 
-class ClassificationHead(block_module.Head):
+class ClassificationHead(head_module.Head):
     """Classification Dense layers.
 
     Use sigmoid and binary crossentropy for binary classification and multi-label
@@ -49,16 +47,11 @@ class ClassificationHead(block_module.Head):
         if not self.metrics:
             self.metrics = ['accuracy']
         self.dropout_rate = dropout_rate
-        self.label_encoder = None
         self.set_loss()
 
     def set_loss(self):
         if not self.num_classes:
             return
-        if self.num_classes <= 1:
-            raise ValueError('Expect the target data for {name} to have '
-                             'at least 2 classes, but got {num_classes}.'
-                             .format(name=self.name, num_classes=self.num_classes))
         if not self.loss:
             if self.num_classes == 2 or self.multi_label:
                 self.loss = 'binary_crossentropy'
@@ -72,20 +65,6 @@ class ClassificationHead(block_module.Head):
             'multi_label': self.multi_label,
             'dropout_rate': self.dropout_rate})
         return config
-
-    def get_state(self):
-        state = super().get_state()
-        state.update({
-            'encoder': encoder.serialize(self.label_encoder),
-            'encoder_state': self.label_encoder.get_state() if self.label_encoder
-            else None})
-        return state
-
-    def set_state(self, state):
-        super().set_state(state)
-        self.label_encoder = encoder.deserialize(state['encoder'])
-        if 'encoder_class' in state:
-            self.label_encoder.set_state(state['encoder_state'])
 
     def build(self, hp, inputs=None):
         if self.num_classes:
@@ -118,50 +97,16 @@ class ClassificationHead(block_module.Head):
             output_node = layers.Softmax(name=self.name)(output_node)
         return output_node
 
-    def _fit(self, y):
-        super()._fit(y)
-        if isinstance(y, tf.data.Dataset):
-            if not self.num_classes:
-                for y in tf.data.Dataset:
-                    shape = y.shape[0]
-                    break
-                if shape == 1:
-                    self.num_classes = 2
-                else:
-                    self.num_classes = shape
-            self.set_loss()
-            return
-        if isinstance(y, pd.DataFrame):
-            y = y.values
-        if isinstance(y, pd.Series):
-            y = y.values.reshape(-1, 1)
-        # Not label.
-        if len(y.flatten()) != len(y):
-            self.num_classes = y.shape[1]
-            self.set_loss()
-            return
-        labels = set(y.flatten())
-        if self.num_classes is None:
-            self.num_classes = len(labels)
-        if self.num_classes == 2:
-            self.label_encoder = encoder.LabelEncoder()
-        elif self.num_classes > 2:
-            self.label_encoder = encoder.OneHotEncoder()
+    def get_adapter(self):
+        return adapters.ClassificationHeadAdapter(name=self.name)
+
+    def config_from_adapter(self, adapter):
+        super().config_from_adapter(adapter)
+        self.num_classes = adapter.num_classes
         self.set_loss()
-        self.label_encoder.fit_with_labels(y)
-
-    def _convert_to_dataset(self, y):
-        if self.label_encoder:
-            y = self.label_encoder.encode(y)
-        return super()._convert_to_dataset(y)
-
-    def postprocess(self, y):
-        if self.label_encoder:
-            y = self.label_encoder.decode(y)
-        return y
 
 
-class RegressionHead(block_module.Head):
+class RegressionHead(head_module.Head):
     """Regression Dense layers.
 
     The targets passing to the head would have to be tf.data.Dataset, np.ndarray,
@@ -221,3 +166,6 @@ class RegressionHead(block_module.Head):
         output_node = layers.Dense(self.output_shape[-1],
                                    name=self.name)(output_node)
         return output_node
+
+    def get_adapter(self):
+        return adapters.RegressionHeadAdapter(name=self.name)
