+++ autokeras/hypermodels/preprocessing.py
@@ -3,8 +3,8 @@ from tensorflow.keras.layers.experimental import preprocessing
 from tensorflow.python.util import nest
 
 from autokeras import keras_layers
+from autokeras.adapters import input_adapter
 from autokeras.engine import block as block_module
-from autokeras.hypermodels import node as node_module
 
 
 class Normalization(block_module.Block):
@@ -194,7 +194,7 @@ class FeatureEncoding(block_module.Block):
         encoding = []
         for column_name in self.column_names:
             column_type = self.column_types[column_name]
-            if column_type == node_module.StructuredDataInput.CATEGORICAL:
+            if column_type == input_adapter.StructuredDataInputAdapter.CATEGORICAL:
                 # TODO: Search to use one-hot or int.
                 encoding.append(keras_layers.INT)
             else:
