+++ autokeras/tasks/__init__.py
@@ -1,0 +1,6 @@
+from autokeras.tasks.image import ImageClassifier
+from autokeras.tasks.image import ImageRegressor
+from autokeras.tasks.structured_data import StructuredDataClassifier
+from autokeras.tasks.structured_data import StructuredDataRegressor
+from autokeras.tasks.text import TextClassifier
+from autokeras.tasks.text import TextRegressor
