+++ autokeras/tuners/__init__.py
@@ -1,0 +1,5 @@
+from autokeras.tuners.bayesian_optimization import BayesianOptimization
+from autokeras.tuners.greedy import Greedy
+from autokeras.tuners.hyperband import Hyperband
+from autokeras.tuners.random_search import RandomSearch
+from autokeras.tuners.task_specific import ImageClassifierTuner
