+++ tests/autokeras/graph_test.py
@@ -1,11 +1,9 @@
-import os
-
 import kerastuner
 import pytest
 from kerastuner.engine import hyperparameters as hp_module
 
 import autokeras as ak
-from autokeras.hypermodels import graph as graph_module
+from autokeras import graph as graph_module
 
 
 @pytest.fixture(scope='module')
@@ -107,11 +105,8 @@ def test_graph_save_load(tmp_dir):
         inputs=[input1, input2],
         outputs=[output1, output2],
         override_hps=[hp_module.Choice('dense_block_1/num_layers', [6], default=6)])
-    path = os.path.join(tmp_dir, 'graph')
-    graph.save(path)
     config = graph.get_config()
     graph = graph_module.Graph.from_config(config)
-    graph.reload(path)
 
     assert len(graph.inputs) == 2
     assert len(graph.outputs) == 2
